Use log component in provision
