+++ stories/components/atoms/log.scss
