+++ stories/components/molecules/index.js
@@ -1,14 +1,14 @@
 import { storiesOf } from '@kadira/storybook';
-import Header from './header';
 import Cluster from './cluster';
-import Service from './service';
-import Log from './log';
+import Header from './header';
 import Host from './host';
+import Provision from './provision';
+import Service from './service';
 
 
 storiesOf('Molecules', module)
   .add('cluster', Cluster)
-  .add('service', Service)
-  .add('log', Log)
   .add('header', Header)
+  .add('provision', Provision)
+  .add('service', Service)
   .add('host', Host);
