+++ stories/components/molecules/provision.scss
@@ -1,0 +1,3 @@
+.label, .item {
+  display: inline;
+}
